Runtime bugfixes. Also greatly increased DPCS VDD transition penalties. The DPCS sampling intervals are now set to 100X the DPCS transition penalty, the objective being to bound performance hit of DPCS transitions to 1 percent.

+++ src/mem/cache/tags/base.hh
@@ -97,7 +97,7 @@ class BaseTags : public ClockedObject
 
   public:
 	unsigned long blockReplacementsInFaultySets; //DPCS: total number of block replacements in faulty sets that have occurred, compared to all block replacements. This is public so that the cache can manipulate this easily for DPCS bookkeeping
-	unsigned long blockReplacementsInFaultySetsRate; //DPCS: rate of block replacements in faulty sets that have occurred, compared to all block replacements. This is public so that the cache can manipulate this easily for DPCS bookkeeping
+	double blockReplacementsInFaultySetsRate; //DPCS: rate of block replacements in faulty sets that have occurred, compared to all block replacements. This is public so that the cache can manipulate this easily for DPCS bookkeeping
 	unsigned long totalBlockReplacements; //DPCS: total number of block replacements that have occurred. This is public so that the cache can manipulate this easily for DPCS bookkeeping.
 
     // Statistics
